requirements.txt
flask
pandas
selenium
webdriver-manager
feedparser
beautifulsoup4
requests
xlsxwriter
python-docx
gunicorn
